lib/group_cpus: optimize outer loop in grp_spread_init_one()
Similarly to the inner loop, in the outer loop we can use for_each_cpu()
macro, and skip CPUs that have been moved.

With this patch, the function becomes O(1), despite that it's a
double-loop.

While here, add a comment why we can't merge outer logic into the inner
loop.

Signed-off-by: Yury Norov <yury.norov@gmail.com>