bitmap: relax find_nth_bit() limitation on return value
Now the function claims to return size, if Nth bit doesn't exist.
This rule is violated in inline case because the fns()
function that is used there doesn't know anything about size of the
bitmap.

So, relax this requirement to returning >= size, and make the
outline implementation a bit cheaper.

All in-tree kernel users of find_nth_bit() are safe against that.

Reported-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Closes: https://lore.kernel.org/all/Zi50cAgR8nZvgLa3@yury-ThinkPad/T/#m6da806a0525e74dcc91f35e5f20766ed4e853e8a
Signed-off-by: Yury Norov <yury.norov@gmail.com>