MAINTAINERS: add BITOPS API record
Bitops API is the very basic, and it's widely used by the kernel.
But corresponding files are not maintained.

Bitmaps actively use bit operations, and big share of bitops material
already moves through the bitmap branch.

I would like to take a closer look to bitops.

This patch creates a BITOPS API record in the MAINTAINERS, and adds
Rasmus as a reviewer, and myself as a maintainer of those files.

Signed-off-by: Yury Norov <yury.norov@gmail.com>