sched: pre-caculate ilog2(TASK_REPORT_MAX)
<linux/sched.h> (indirectly, via cpumask.h) path includes the ilog2.h
header to calculate ilog2(TASK_REPORT_MAX).

The following patches drops sched.h dependency on cpumask.h, and
to have a successful build the header has to be included explicitly.

sched.h is frequently included header, and it's better to keep the
dependency list as small as possible. So, instead of including ilog2.h
for a single BUILD_BUG_ON() check of ilog2(TASK_REPORT_MAX), the same
check may be implemented by taking exponent of the other part of
equation.

Signed-off-by: Yury Norov <yury.norov@gmail.com>