sched: drop dependency on cpumask.h
sched.h needs cpumask.h mostly for types declaration. Now that we have
cpumask_types.h, which is a significantly smaller header, we can rely on
it.

The only exception is UP fallback for set_cpus_allowed_ptr(). The
function needs to test bit #0 in a @new_mask, which can be trivially
opencoded.

Signed-off-by: Yury Norov <yury.norov@gmail.com>