cpumask: optimize cpumask_any_and_but()
If (mask1 & mask2) is empty, the cpumask_first_and() returns a value
greater than nr_cpu_ids. This should be enough for the function to
return. But current implementation ignores it and calls find_next_bit(),
which is useless.

Instead of adding that check, we can align the cpumask_any_and_but()
implementation with cpumask_any_but() by using a for_each_cpu_and()
loop.