bitops: add compile-time test for ror() and rol()
If parameters for the functions are pased at compile time, the compiler
must calculate the result at comlile time, as well.

Now that architectures introduce accelerated implementations for bit
rotation, we must make sure that they don't break compile-time
evaluation.

This patch adds a test for it, similarly to test_bitmap_const_eval().