loongarch: rework kvm_send_pv_ipi()
The function in fact traverses a "bitmap" stored in GPR args
A1 and A2, but does it in a non-obvious way, by creating a single-word
bitmap twice.

This patch switches the function to create a single 2-word bitmap, and
also employs for_each_set_bit() macro, as it helps to drop most of
housekeeping code.